feat: add commentaries & end work

+++ internal/domain/models/aiservice.go
@@ -4,11 +4,11 @@ type AIService struct {
 	UID         string
 	Title       string
 	Description string
-	Price       int
+	Price       float64
 }
 
 type AIServiceCreate struct {
 	Title       string
 	Description *string
-	Price       int
+	Price       float64
 }
